Add shipment_id to the loads and shipments APIs (#96)
Right now shipment external_id is used for both the machine-readable
identifier and the human-readable identifier. This splits them into
shipment_id and external_id similar to loads.

We're adding shipment_id and load_id to the areas that link them back
and forth so we can create the shipment or loads with the right
identifiers.